requirements.txt
# Core Data Science & ML
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning & NLP
torch==2.0.1
transformers==4.30.2
sentence-transformers==2.2.2

# Time Series Analysis
prophet==1.1.4
statsmodels==0.14.0
pmdarima==2.0.3

# Web Scraping & APIs
beautifulsoup4==4.12.2
selenium==4.11.2
requests==2.31.0
aiohttp==3.8.5
scrapy==2.9.0

# Real-Time Processing
apache-kafka-python==2.0.2
redis==4.6.0
celery==5.3.1

# Dashboard & Visualization
streamlit==1.25.0
plotly==5.15.0
altair==5.0.1
pydeck==0.8.0

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
pymongo==4.4.1

# NLP & Text Processing
nltk==3.8.1
spacy==3.6.0
langdetect==1.0.9
textblob==0.17.1

# Scheduling & Background Tasks
APScheduler==3.10.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.0
tqdm==4.65.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# API Development (optional)
fastapi==0.100.0
uvicorn==0.23.1

# Cloud & Deployment
gunicorn==21.2.0
docker==6.1.3